# Replace ChaCha20Poly1305 implementation