Improve AEAD speed with slightly faster poly1305 implementation.
Avoid memory allocations whenever possible. (AEAD)
But currently missing AVX2 support.

BenchmarkSeal64B-8  &nbsp;        &nbsp;  &nbsp;40.97 MB/s
BenchmarkSeal1K-8 &nbsp; &nbsp;  &nbsp; 183.82 MB/s
BenchmarkSeal64K-8  &nbsp;     &nbsp;406.37 MB/s
BenchmarkOpen64B-8    &nbsp; &nbsp; 45.79 MB/s
BenchmarkOpen1K-8  &nbsp;    &nbsp;   181.14 MB/s
BenchmarkOpen64K-8   &nbsp;  417.22 MB/s