@aead 2016-07-18
The Travis-CI build seems to fail because of a network error...
> unable to access 'https://go.googlesource.com/tools/