@lucas-clemente 2016-07-18
Thanks for this PR! I restarted the travis build and it worked, so I'm going ahead with merging this now.

I've had a quick look at your implementation, and it indeed looks better with regard to allocations. It also allows us to use in-place encryption, which will probably be done in our current milestone (see #217).

However, I'm not sure if you're aware that we're only using AES-GCM right now. We switched since it is around an order of magnitude faster (with the current implementations, see below). Due to a lack of time I didn't build proper negotiation for chacha yet (the client gets to decide, see #201), but I will do this soon-ish.

Just pasting my benchmark results for future reference (of course they don't cover GC pressure):

```
Old implementation:
BenchmarkChaCha20Poly1305Seal1K-8   	  500000	      2804 ns/op	 365.13 MB/s
BenchmarkChaCha20Poly1305Open1K-8   	  500000	      2808 ns/op	 364.60 MB/s

New implementation:
BenchmarkSeal1K-8            	  500000	      2745 ns/op	 373.03 MB/s
BenchmarkOpen1K-8            	  500000	      2732 ns/op	 374.76 MB/s

AES-GCM:
BenchmarkAESGCMSeal1K-8   	 3000000	       466 ns/op	2196.10 MB/s
BenchmarkAESGCMOpen1K-8   	 3000000	       440 ns/op	2326.29 MB/s
```