@aead 2016-07-18
Thanks for merging!  

It's quite hard to beat AES-GCM on amd64 with AES-NI - Maybe AVX2 has a chance (but both poly1305 and chacha20 must use it).
I'm working on that as soon as I have some time left.  

Thanks for the benchmark reference - quite helpful!