@aead 2016-07-21
Significant performance improvement for poly1305 through amd64 assembly.
But still not comparable with AES-NI:

BenchmarkSeal64B-8 &nbsp; &nbsp; &nbsp; 57.89 MB/s 
BenchmarkSeal1K-8 &nbsp; &nbsp; &nbsp; 298.51 MB/s 
BenchmarkSeal64K-8 &nbsp; &nbsp;  688.23 MB/s 
BenchmarkOpen64B-8 &nbsp; &nbsp;  47.62 MB/s 
BenchmarkOpen1K-8 &nbsp; &nbsp; 302.07 MB/s 
BenchmarkOpen64K-8 &nbsp; 678.04 MB/s