@lucas-clemente 2016-07-21
Very cool work! :+1: On my machine it's around 420, up from 370.

```
BenchmarkSeal1K-8            	 1000000	      2435 ns/op	 420.48 MB/s
BenchmarkOpen1K-8            	  500000	      2431 ns/op	 421.14 MB/s
```

Once QUIC gets adopted by mobile browsers (who like choosing chacha over AES) this will become really useful :)