@aead 2016-07-25
Thanks :smile:
I've some working (experimental) AVX2 code for ChaCha20 giving a great speedup again, but only go-1.7 will support AVX2 asm instructions (VPSHUFD etc.). The current go 1.7beta-3 have still some problems with AVX2 instructions - so readable asm-code requires the final go 1.7 release. 

The main problem for you now is the compilation. If I add the AVX2 code to the repo (as soon as the release is available), only the 1.7 builds will pass ( <= 1.6.3 will fail). As far as I know Go does not support something like "version depended compilation", so I cannot fix this...
Do you want to keep support for older versions of Go ( <= 1.6.3) ?  

Currently i think i'll create a AVX2 branch for chacha20 and as far as possible (without breaking other's code) make it the master...

Some benchmarks for AVX2 code on a Skylake i7-6500U:
Without AVX2:
- BenchmarkSeal1K-4 565.44 MB/s
- BenchmarkOpen1K-4 559.03 MB/s

With AVX2:
+ BenchmarkSeal1K-4 791.02 MB/s
+ BenchmarkOpen1K-4 779.34 MB/s