@lucas-clemente 2016-07-25
I would like to support keep support for 1.6 for the near future.

However, I think it's possible to use Go's build tags for that purpose, see https://golang.org/pkg/go/build/#hdr-Build_Constraints.

In short, if you put `// +build go1.7` into the AVX2 files, they will only be compiled on go 1.7 or up. Similarly for the old implementation.