@aead 2016-07-25
Thanks for the info - perfect :+1:
Okay, AVX2 will be supported as soon as 1.7 is stable...