@aead 2016-07-26
SSSE3 support available: aead/chacha20@b2542acc9b7dab45d6ac7f7ecf4fce58a06eb79b
Performance improvement about 8-10% on my machine  
Further fixed a critical bug detected by [v2ray](https://github.com/v2ray/v2ray-core/pull/215#issuecomment-235071183) See  aead/chacha20@b9b9fe0318a29cfdf8024f4e2e92a735c23529a3 
Dependency update necessary...